+++ app/src/main/java/org/simple/clinic/editpatient/EditPatientUi.kt
@@ -7,7 +7,7 @@ import org.simple.clinic.scanid.OpenedFrom
 import org.simple.clinic.widgets.ageanddateofbirth.DateOfBirthAndAgeVisibility
 
 interface EditPatientUi {
-  fun displayBpPassports(identifiers: List<String>)
+  fun displayBpPassports(bpPassports: List<BPPassportListItem>)
   fun setPatientName(name: String)
   fun setPatientPhoneNumber(number: String)
   fun setColonyOrVillage(colonyOrVillage: String)